requirements.txt
# Core Dependencies
python>=3.9
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.0

# LLM Libraries
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
torch>=2.0.0
sentencepiece>=0.1.99

# Optional (for future API support)
openai>=1.0.0
anthropic>=0.7.0

# Utilities
tqdm>=4.65.0
jsonschema>=4.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
